# backend/src/swagger/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BillFlow ERP API
  version: 1.0.0
  description: |
    GST Billing & Invoice Management System for Manufacturing Companies
    
    ## Overview
    BillFlow ERP provides a comprehensive API for managing invoices, products, customers, and GST compliance.
    
    ## Authentication
    All endpoints (except health) require JWT authentication via Bearer token.
    
    ## Rate Limiting
    API requests are rate-limited to 100 requests per 15 minutes per IP.
    
    ## Versioning
    API is versioned via URL path: `/api/v1`
    
    ## Response Format
    All responses follow a consistent format:
    ```json
    {
      "success": true,
      "message": "Success",
      "data": {}
    }
    ```
    
    ## Error Format
    ```json
    {
      "success": false,
      "message": "Error message",
      "errors": [],
      "timestamp": "2024-01-01T00:00:00.000Z",
      "path": "/api/v1/resource",
      "requestId": "uuid"
    }
    ```
  contact:
    name: BillFlow Team
    email: support@billflow.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.billflow.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Success
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation Failed
        errors:
          type: array
          items:
            type: object
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        requestId:
          type: string
...
security:
  - bearerAuth: []
...
paths:
...
  /storage/upload/image:
    post:
      tags: [Storage]
      summary: Upload an image file
      description: |
        Upload an image file to storage (ADMIN only).
        
        **Supported formats:** JPEG, PNG, WebP, SVG
        **Maximum file size:** 5 MB
        **Allowed folders:** company, products, signatures
        
        The file will be renamed to `<uuid>_<timestamp>.<extension>` format
        and stored in the specified folder.
      operationId: uploadImage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file to upload
      parameters:
        - name: folder
          in: query
          description: Target folder for the upload
          required: false
          schema:
            type: string
            enum: [company, products, signatures]
            default: products
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UploadedFile'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /storage/upload/signature:
    post:
      tags: [Storage]
      summary: Upload a signature file
      description: |
        Upload a signature image to storage (ADMIN only).
        
        **Supported formats:** JPEG, PNG, WebP, SVG
        **Maximum file size:** 5 MB
        **Folder:** signatures (fixed)
        
        The file will be renamed to `<uuid>_<timestamp>.<extension>` format
        and stored in the signatures folder.
      operationId: uploadSignature
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Signature image file to upload
      responses:
        '201':
          description: Signature uploaded successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UploadedFile'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /storage/{path}:
    delete:
      tags: [Storage]
      summary: Delete a file
      description: |
        Delete a file from storage by its path (ADMIN only).
        
        The path should be the relative path returned from upload (e.g., `products/3f21d88d_1754300000.webp`).
      operationId: deleteFile
      security:
        - bearerAuth: []
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
          description: File path to delete
      responses:
        '200':
          description: File deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Path parameter is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /storage/url/{path}:
    get:
      tags: [Storage]
      summary: Get public URL for a file
      description: |
        Get the public URL for a file in storage.
        
        The path should be the relative path returned from upload (e.g., `products/3f21d88d_1754300000.webp`).
      operationId: getFileUrl
      security:
        - bearerAuth: []
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
          description: File path to get URL for
      responses:
        '200':
          description: File URL retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          url:
                            type: string
                            format: uri
                            example: "https://example.com/storage/products/3f21d88d_1754300000.webp"
        '400':
          description: Path parameter is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /invoices/next-number:
    get:
      tags: [Invoices]
      summary: Get next invoice number
      description: Returns the next invoice number that will be assigned (ADMIN, ACCOUNTANT, SALES)
      operationId: getNextInvoiceNumber
      responses:
        '200':
          description: Next invoice number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...